more gchart update
git-svn-id: https://pyjamas.svn.sourceforge.net/svnroot/pyjamas/trunk@1594 7a2bd370-bda8-463c-979e-2900ccfb811e

+++ library/pyjamas/chart/GChartWidgets.py
@@ -171,8 +171,8 @@ class PartitionedAbsolutePanel (Composite):
 
     # makes the subpanel containing the widget the selected one.
     def selectSubPanel(self, iWidget):
-        if self.iSubPanel != iWidget/WIDGETS_PER_PANEL:
-            self.iSubPanel = iWidget/WIDGETS_PER_PANEL
+        if self.iSubPanel != int(iWidget/WIDGETS_PER_PANEL):
+            self.iSubPanel = int(iWidget/WIDGETS_PER_PANEL)
             self.subPanel = self.root.getWidget(self.iSubPanel)
